ecomSys_v2/mobile/requirements.txt
Django~=4.1.13
PyJWT~=2.8.0
requests~=2.31.0
djangorestframework~=3.14.0
djongo~=1.3.6
python-magic~=0.4.27
Cerberus~=1.3.5
pymongo~=3.12.3